[arthur] Fix error initializing threads on daemon mode
Some threads like the scheduler thread or the writer thread
were initialized before daemonizing (forking) the script.
This is a bug beacuse threads are not shared after a fork() call.

To fix it, threads must be started after the cherrypy plugin
for daemonizing is run. This plugin waits for the 'start'
event to do its fork call. Arthur server is subscribed to the
'start' events too but with a priority lower than the daemon
plugin. That is how the threads will be started after the daemon
initialization.