[jobs] Support resuming of jobs
When a job fails, it will run again trying to resume from
the point it failed. If the limit of max retries is reached,
the job will be cancelled and sent to the queue of failed
jobs.

To track the number of times a job has been resumed,
JobResult includes a new attribute 'nresumed'.