[scheduler] Refactorize scheduler to accept tasks
The scheduler accepts tasks as input, scheduling them into
Perceval jobs. The new scheduler supports to cancel tasks, too.

It runs a thread that will listen for finished jobs, when the
job is successful, it will be scheduled again to fetch new data;
if the result is an error, the job will be inmediately canceled.