[tests] Add fakeredis to unit tests
The unit tests need a connection with a Redis database.
With this change, the real server is replaced by
fakeredis library to emulate the use of a server.

As ArthurWorker class inherits from rq.Worker,
the class MockArthurWorker has been created to avoid
fork call when a job is executed. The call to fork makes
fakeredis could not be used because it can only work with
threads (it uses shared memory to send messages between
workers). This error has been solved using the SimpleWorker
class from rq module.