[jobs] Fix 'from_date' and 'offset' reset error
So far, when a job runs in the 'update' queue and no new items
are fetched, either 'from_date' or 'offset' values are reset
to the starting values defined for the parent task (in backend_args).
This is an error because jobs in 'update' must start from the
point where the previous job ended.