[events] Extract job listener to create a job events listener
When job was completed, the worker sent a message to notify
about whether the job finished correctly or it failed.
So far, a thread, running inside the scheduler, was listening
for those messages and scheduling new jobs. This creates
a coupling between classes and responsibilities. A listener
should only do that, and should live inside the scheduler.

This commit re-factorizes the scheduler and the listener. A new
class named `JobEventsListener`, defined in `events` module,
has been extracted from the scheduler. This class only does
the same that the old listener did. Further versions will
manage job events.

Signed-off-by: Santiago Dueñas <sduenas@bitergia.com>