[events] Add basic job events
A new class JobEvent is introduced in this commit to send
job notifications using events. There two event types defined
so far: COMPLETED and FAILURE. Events contain information
such as the timestamp when it was generated, the job that
created it and a variable payload to store specific data.

Events can be serialized and deserialized to send them
though Redis pubsub channels.

Signed-off-by: Santiago Dueñas <sduenas@bitergia.com>