[scheduler] Delegate queue selection to the task scheduler
With this change 'TaskScheduler' class will be de one
responsible to determine which tasks go to which
queues.

It was needed to add a 'age' attribute to 'Task' class
to determine if the task was re-scheduled. This value
will increase each time the task is enqueued again.
For those kind of tasks, the default query for their
jobs is Q_UPDATING_JOBS.

Signed-off-by: Santiago Dueñas <sduenas@bitergia.com>