[scheduler] Refactor event handlers
Event handlers were integrated within Scheduler class.
Among coupling problems, this made impossible to test
whether events were correctly handled or not because
they only worked when the listener was running.
With this patch we fix those problems.

To make it work, it was also necessary to modify the
payload of a failed job. The payload now includes the
'task_id' and the exception of the error.

Signed-off-by: Santiago Dueñas <sduenas@bitergia.com>