[writers] Ignore Elasticsearch certificate verification
This patch sets requests to ignore certificate validation
of the Elasticsearch server where items will be written.

Take into account this writer is used with testing purposes
only. It should not be used in production. As testing is
the normal case, it is not worthy to have a certificate
for a local server.

Signed-off-by: Santiago Dueñas <sduenas@bitergia.com>