[jobs] Generate items using BackendItemsGenerator
The class 'BackendItemsGenerator' from 'perceval' package
provides an easy way to generate items and to obtain a
summary with the results while fetching process is running.

This commit replaces the call to the alternative methods
'fetch' and 'fetch_from_archive' that are deprecated and
removed soon.

This is the first steep needed to use the summary generated
by Perceval.

Signed-off-by: Santiago Dueñas <sduenas@bitergia.com>