[jobs] Refactor 'metadata' decorator
The decorator 'metadata' has been refactored and moved to
the class PercevalJob. Now it is a private method called
when items are generated inside the method 'run'.

This change was needed because the method 'execute' was
decorated by this function but it was replaced for a method
that creates an items generator.

Signed-off-by: Santiago Dueñas <sduenas@bitergia.com>