[jobs] Store logs of each job in their "meta" field
In order to save the logs of each job, this commit adds the
functionality of storing them in the "meta" field of the job, that
will be saved in Redis.

This commit includes a new log handler for the arthur, perceval
and rq.worker loggers. These loggers have information about the
job execution, so the logs are filtered by the job id before
save them.

Signed-off-by: David Moreno <dmoreno@bitergia.com>