[jobs] Replace job result object by Perceval summary
Some information related to the progress of a job was
stored within JobResult. Since Perceval provides 'Summary'
objects for each fetch process, this information is
redundant.

This commit replaces this information by the one provided
by Perceval, which is more accurated and completed. For
example, it stores the number of items generated and skipped;
the UUID of last item generated; max and min dates of the
items and so on.

Signed-off-by: Santiago Dueñas <sduenas@bitergia.com>