[tasks] Add has_resuming() method to task
So far jobs store whether they can be resumed when they
fail. Tasks will be responsible to reschedule these jobs
so they will need to know if it is possible to do it.
This is usally defined by the Perceval backend in use.

This commit adds a method that allows to determine if
it is possible to restart a task from the point it failed.

Signed-off-by: Santiago Dueñas <sduenas@bitergia.com>