[scheduler] Remove 'arthur' prefix and task id from job ids schema
The new schema for job ids consists on a single hash, generated
using the `uuid` module. The former schema was:
`arthur-<task_id>-hash`.

The `arthur-` prefix was set in the past to identify jobs coming
from different sources (not from `arthur`). By removing this
prefix the job ids are now simpler, as the origin of the job
could be included as meta-data instead of a part of the job id.

The `task_id` parameter came from the task triggering the job.
By removing this section, we are avoiding future problems as this
field could contain whitespaces or other unusual characters.

Signed-off-by: Miguel Ángel Fernández <mafesan@bitergia.com>