[scheduler] Allow to reset task counters
A new parameter 'reset' has been added to 'schedule_task()`
method. When this parameter is set some of the counters
of a task will be reset to their initial value. This will
be useful in the case if re-scheduling failed tasks due
to values like `max_age` can stop running jobs again.

Signed-off-by: Santiago Dueñas <sduenas@bitergia.com>