[tasks] Add JobData tuple to store jobs in the task job list
In order to return the job_number field and the job_id identifier,
this tuple is created with the goal of store more than the
identifier in the task job list. Thus, when the server querys the
tasks, it will return a list of JobData objects that contains the
job_number field and the job_id identifier.

This commit affects to the scheduler and server as well.

Signed-off-by: David Moreno <dmoreno@bitergia.com>